fix module issue with web build
