+++ lib/react_native/index.js
@@ -23,17 +23,17 @@ var buffer_1 = require("buffer");
 var pubnub_common_1 = __importDefault(require("../core/pubnub-common"));
 var networking_1 = __importDefault(require("../networking"));
 var base64_codec_1 = require("../core/components/base64_codec");
+var stringify_buffer_keys_1 = require("../core/components/stringify_buffer_keys");
 var common_1 = __importDefault(require("../cbor/common"));
 var web_node_1 = require("../networking/modules/web-node");
 var react_native_1 = require("../networking/modules/react_native");
 var react_native_2 = __importDefault(require("../file/modules/react-native"));
-var web_1 = require("../web");
 global.Buffer = global.Buffer || buffer_1.Buffer;
 var default_1 = /** @class */ (function (_super) {
     __extends(default_1, _super);
     function default_1(setup) {
         var _this = this;
-        setup.cbor = new common_1.default(function (arrayBuffer) { return (0, web_1.stringifyBufferKeys)(cbor_js_1.default.decode(arrayBuffer)); }, base64_codec_1.decode);
+        setup.cbor = new common_1.default(function (arrayBuffer) { return (0, stringify_buffer_keys_1.stringifyBufferKeys)(cbor_js_1.default.decode(arrayBuffer)); }, base64_codec_1.decode);
         setup.PubNubFile = react_native_2.default;
         setup.networking = new networking_1.default({
             del: web_node_1.del,
