+++ lib/web/index.js
@@ -20,11 +20,11 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.stringifyBufferKeys = void 0;
 var cbor_js_1 = __importDefault(require("cbor-js"));
 var pubnub_common_1 = __importDefault(require("../core/pubnub-common"));
 var networking_1 = __importDefault(require("../networking"));
 var base64_codec_1 = require("../core/components/base64_codec");
+var stringify_buffer_keys_1 = require("../core/components/stringify_buffer_keys");
 var common_1 = __importDefault(require("../cbor/common"));
 var web_node_1 = require("../networking/modules/web-node");
 var web_1 = __importDefault(require("../crypto/modules/web"));
@@ -37,33 +37,6 @@ function sendBeacon(url) {
         return false;
     }
 }
-function stringifyBufferKeys(obj) {
-    var isObject = function (value) { return value && typeof value === 'object' && value.constructor === Object; };
-    var isString = function (value) { return typeof value === 'string' || value instanceof String; };
-    var isNumber = function (value) { return typeof value === 'number' && isFinite(value); };
-    if (!isObject(obj)) {
-        return obj;
-    }
-    var normalizedObject = {};
-    Object.keys(obj).forEach(function (key) {
-        var keyIsString = isString(key);
-        var stringifiedKey = key;
-        var value = obj[key];
-        if (Array.isArray(key) || (keyIsString && key.indexOf(',') >= 0)) {
-            var bytes = keyIsString ? key.split(',') : key;
-            stringifiedKey = bytes.reduce(function (string, byte) {
-                string += String.fromCharCode(byte);
-                return string;
-            }, '');
-        }
-        else if (isNumber(key) || (keyIsString && !isNaN(key))) {
-            stringifiedKey = String.fromCharCode(keyIsString ? parseInt(key, 10) : 10);
-        }
-        normalizedObject[stringifiedKey] = isObject(value) ? stringifyBufferKeys(value) : value;
-    });
-    return normalizedObject;
-}
-exports.stringifyBufferKeys = stringifyBufferKeys;
 var default_1 = /** @class */ (function (_super) {
     __extends(default_1, _super);
     function default_1(setup) {
@@ -80,7 +53,7 @@ var default_1 = /** @class */ (function (_super) {
             getfile: web_node_1.getfile,
             postfile: web_node_1.postfile,
         });
-        setup.cbor = new common_1.default(function (arrayBuffer) { return stringifyBufferKeys(cbor_js_1.default.decode(arrayBuffer)); }, base64_codec_1.decode);
+        setup.cbor = new common_1.default(function (arrayBuffer) { return (0, stringify_buffer_keys_1.stringifyBufferKeys)(cbor_js_1.default.decode(arrayBuffer)); }, base64_codec_1.decode);
         setup.PubNubFile = web_2.default;
         setup.cryptography = new web_1.default();
         _this = _super.call(this, setup) || this;
