+++ src/react_native/index.js
@@ -3,6 +3,7 @@ import { Buffer } from 'buffer';
 import PubNubCore from '../core/pubnub-common';
 import Networking from '../networking';
 import { decode } from '../core/components/base64_codec';
+import { stringifyBufferKeys } from '../core/components/stringify_buffer_keys';
 import Cbor from '../cbor/common';
 import { del, get, post, patch } from '../networking/modules/web-node';
 import { getfile, postfile } from '../networking/modules/react_native';
