+++ src/web/index.js
@@ -5,6 +5,7 @@ import CborReader from 'cbor-js';
 import PubNubCore from '../core/pubnub-common';
 import Networking from '../networking';
 import { decode } from '../core/components/base64_codec';
+import { stringifyBufferKeys } from '../core/components/stringify_buffer_keys';
 import Cbor from '../cbor/common';
 import { del, get, post, patch, getfile, postfile } from '../networking/modules/web-node';
 
@@ -17,39 +18,6 @@ function sendBeacon(url) {
   } else {
     return false;
   }
-}
-
-export function stringifyBufferKeys(obj) {
-  const isObject = (value) => value && typeof value === 'object' && value.constructor === Object;
-  const isString = (value) => typeof value === 'string' || value instanceof String;
-  const isNumber = (value) => typeof value === 'number' && isFinite(value);
-
-  if (!isObject(obj)) {
-    return obj;
-  }
-
-  const normalizedObject = {};
-
-  Object.keys(obj).forEach((key) => {
-    const keyIsString = isString(key);
-    let stringifiedKey = key;
-    const value = obj[key];
-
-    if (Array.isArray(key) || (keyIsString && key.indexOf(',') >= 0)) {
-      const bytes = keyIsString ? key.split(',') : key;
-
-      stringifiedKey = bytes.reduce((string, byte) => {
-        string += String.fromCharCode(byte);
-        return string;
-      }, '');
-    } else if (isNumber(key) || (keyIsString && !isNaN(key))) {
-      stringifiedKey = String.fromCharCode(keyIsString ? parseInt(key, 10) : 10);
-    }
-
-    normalizedObject[stringifiedKey] = isObject(value) ? stringifyBufferKeys(value) : value;
-  });
-
-  return normalizedObject;
 }
 
 export default class extends PubNubCore {
